new packages for core new scripts for building the core finished off help items for commands other changes to the core to work with the new formats

+++ tools/lib/def.php
@@ -1,6 +1,6 @@
 <?php
 
-abstract class def {
+abstract class Def {
 
 	//main data access
 	public $data     = null;
@@ -10,10 +10,12 @@ abstract class def {
 	protected $dataMD5  = null;
 	protected $dataRaw  = null;
 	
+	protected $container_var = 'def';
+	
 	//io handling
-	const READONLY  = true;
-	const READWRITE = false;
-	private $iostate = self::READONLY; // read-only by default
+	const READONLY  = 1;
+	const READWRITE = 0;
+	public $iostate = self::READONLY; // read-only by default
 	
 	function __destruct(){
 		$this->write();
@@ -47,15 +49,15 @@ abstract class def {
 	}
 
 	protected function read(){
-		if((!file_exists($this->filename)) || (($filesize = filesize($this->filename))===0)) $this->write();
-		$md5 = md5_file($this->filename);
+		if((!file_exists($this->filename)) || (($filesize = filesize($this->filename))===0)) $this->sync();
+		$md5 = md5($this->dataRaw);
 		if($this->dataMD5 != $md5){
 			$fh = fopen($this->filename,'r');
 			if(is_resource($fh) && ($def_data = fread($fh,$filesize))){
 				$this->dataRaw = $def_data;
 				$proc = proc_open('php',array(0=>array('pipe','r'),1=>array('pipe','w')),$p);
 				if(is_resource($proc)){
-					fwrite($p[0],$def_data . 'print(serialize($pkgdef));');
+					fwrite($p[0],$def_data . 'print(serialize($'.$this->container_var.'));');
 					fclose($p[0]);
 					$this->data = unserialize(stream_get_contents($p[1]));
 					fclose($p[1]);
@@ -70,13 +72,14 @@ abstract class def {
 
 	protected function sync(){
 		$this->dataSanitizer();
+		$container_var = $this->container_var;
 		$this->dataRaw = <<<'HEADER'
 <?php
 // THIS IS A VOLATILE FILE AND MAY BE REGENERATED AUTOMATICALLY
 
-$pkgdef = array();
 
 HEADER;
+		$this->dataRaw .= '$'.$this->container_var.' = array();'."\n";
 		$this->dataRaw .= self::dumpDef($this->data);
 		$this->dataRaw .= PHP_EOL;
 		$this->dataMD5 = md5($this->dataRaw);
@@ -85,7 +88,7 @@ HEADER;
 	
 	protected function write(){
 		$this->sync();
-//		if($this->iostate === self::READONLY) return $this;
+		if($this->iostate === self::READONLY) return $this;
 		if(file_exists($this->filename) && (filesize($this->filename) !== 0))
 			run('mv '.$this->filename.' '.$this->filename.'.bak'); // why not save the previous version
 		@mkdir(dirname($this->filename),0755,true);
@@ -108,11 +111,11 @@ HEADER;
 	//	ksort($def);
 		foreach($def as $def_key => $def_val){
 			if(is_array($def_val)){
-				$output .= sprintf('$pkgdef%s[\'%s\'] = array();'.PHP_EOL,$parent_nodes,$def_key);
+				$output .= sprintf('$%s%s[\'%s\'] = array();'.PHP_EOL,$this->container_var,$parent_nodes,$def_key);
 				$output .= self::dumpDef($def_val,array_merge($parents,array($def_key)));
 			} else {
 				if(!is_numeric($def_key)) $def_key = "'".$def_key."'";
-				$output .= sprintf('$pkgdef%s[%s] = \'%s\';'.PHP_EOL,$parent_nodes,$def_key,$def_val);
+				$output .= sprintf('$%s%s[%s] = \'%s\';'.PHP_EOL,$this->container_var,$parent_nodes,$def_key,$def_val);
 			}
 		}
 		return $output;
